Assignment 4.1
Create a Servlet that displays a form when the doGet method is invoked. The form will contain a post action that directs the form post back to the same servlet, which in the doPost method will append the form data to a random access file. After the form data has been appended to the file, respond back with the complete contents of the data file. The form must contain a minimum of three input fields. The grade for this assignment will be based both on the functionality of the servlet and the appearance of the form post results. Name your servlet <yourName>FormPost and name the random access data file c:\temp\<yourname>week4.dat. You must locate the file in the c:\temp directory to receive credit. This is so I can run and grade it on my computer without modifying your source code. Attach your servlet source code file and random access data file to this assignment in one zipped file.