Assignment 3.1
Create a servlet that reads an ascii file and sends the output to the browser that invokes the servlet. Name your ASCII file c:\temp\servlet1.dat. You must locate it in the c:\temp directory to receive credit. This is so I can run and grade it on my computer without modifying your source code. Hint: use the file I/O examples we covered previously and use absolute file paths. A good file to read is the source code to your program! Attach your servlet source code file and data file to this assignment in one zipped file. Do not copy (cut and paste) any example code, create your own code and use the examples as a guide.